requirements.txt
beautifulsoup4
feedparser
hitherdither @ git+https://www.github.com/hbldh/hitherdither
irc
jaraco.util
audioread
mutagen
numpy
opencv_python
geopy
youtube-dl
pafy @ git+https://github.com/Cupcakus/pafy
Pillow>=8.0.0
python_weather
requests>=2.26.0
streamlink
tinydb
wikipedia
Wikipedia_API
crc